requirements.txt
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
beautifulsoup4==4.13.3
build==1.2.2.post1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Cython==3.0.12
decorator==5.2.1
distro==1.9.0
elevenlabs==1.56.0
filelock==3.18.0
fire==0.7.0
fonttools==4.57.0
fsspec==2025.3.2
greenlet==3.1.1
groq==0.22.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-core==0.3.51
langchain-groq==0.3.2
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.8
langsmith==0.3.30
lazy_loader==0.4
lmdb==1.6.2
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
orjson==3.10.16
packaging==24.2
paddleocr==2.10.0
paddlepaddle==3.0.0
pillow==11.2.1
protobuf==6.30.2
pyclipper==1.3.0.post6
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyproject_hooks==1.2.0
python-docx==1.1.2
python-dotenv==1.1.0
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.2
setuptools==78.1.0
shapely==2.1.0
shellingham==1.5.4
simsimd==6.2.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
stringzilla==3.12.4
sympy==1.13.1
tenacity==9.1.2
termcolor==3.0.1
threadpoolctl==3.6.0
tifffile==2025.3.30
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1
wheel==0.45.1
zstandard==0.23.0
